+++ main/src/ca/uwaterloo/flix/api/lsp/LanguageServer.scala
@@ -17,10 +17,10 @@ package ca.uwaterloo.flix.api.lsp
 
 import ca.uwaterloo.flix.api.lsp.provider._
 import ca.uwaterloo.flix.api.{Flix, Version}
-import ca.uwaterloo.flix.language.ast.Ast.Source
 import ca.uwaterloo.flix.language.ast.TypedAst.Root
-import ca.uwaterloo.flix.language.ast.{Ast, SourceLocation, Symbol}
+import ca.uwaterloo.flix.language.ast.{SourceLocation, Symbol}
 import ca.uwaterloo.flix.language.debug._
+import ca.uwaterloo.flix.language.phase.extra.CodeHinter
 import ca.uwaterloo.flix.util.Result.{Err, Ok}
 import ca.uwaterloo.flix.util.Validation.{Failure, Success}
 import ca.uwaterloo.flix.util.vt.TerminalContext
@@ -178,6 +178,7 @@ class LanguageServer(port: Int) extends WebSocketServer(new InetSocketAddress("l
       case JString("lsp/goto") => Request.parseGoto(json)
       case JString("lsp/rename") => Request.parseRename(json)
       case JString("lsp/documentSymbols") => Request.parseDocumentSymbols(json)
+      case JString("lsp/workspaceSymbols") => Request.parseWorkspaceSymbols(json)
       case JString("lsp/uses") => Request.parseUses(json)
       case JString("lsp/semanticTokens") => Request.parseSemanticTokens(json)
 
@@ -247,6 +248,9 @@ class LanguageServer(port: Int) extends WebSocketServer(new InetSocketAddress("l
     case Request.DocumentSymbols(id, uri) =>
       ("id" -> id) ~ ("status" -> "success") ~ ("result" -> SymbolProvider.processDocumentSymbols(uri)(root).map(_.toJSON))
 
+    case Request.WorkspaceSymbols(id, query) =>
+      ("id" -> id) ~ ("status" -> "success") ~ ("result" -> SymbolProvider.processWorkspaceSymbols(query)(root).map(_.toJSON))
+
     case Request.Uses(id, uri, pos) =>
       ("id" -> id) ~ FindReferencesProvider.findRefs(uri, pos)(index, root)
 
@@ -287,8 +291,18 @@ class LanguageServer(port: Int) extends WebSocketServer(new InetSocketAddress("l
           // Compute elapsed time.
           val e = System.nanoTime() - t
 
-          // Send back a status message.
-          ("id" -> requestId) ~ ("status" -> "success") ~ ("time" -> e)
+          // Compute Code Quality hints.
+          val hints = CodeHinter.run(root)(flix)
+
+          hints match {
+            case Success(_) =>
+              // Case 1: No code hints.
+              ("id" -> requestId) ~ ("status" -> "success") ~ ("time" -> e)
+            case Failure(errors) =>
+              // Case 2: Code hints are available.
+              val results = PublishDiagnosticsParams.from(errors)
+              ("id" -> requestId) ~ ("status" -> "failure") ~ ("result" -> results.map(_.toJSON))
+          }
 
         case Failure(errors) =>
           // Case 2: Compilation failed. Send back the error messages.